requirements.txt
aiohttp==3.8.3
aiohttp-jinja2==1.5
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
flower==1.2.0
frozenlist==1.3.3
humanize==4.4.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
multidict==6.0.4
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pytz==2022.7.1
redis==4.4.2
six==1.16.0
tornado==6.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
